+++ libs/sdk-core/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Improvements
 
 - Add `show_consumption` option on `/ask`
+- Support importing and exporting Retrieval Agents
 
 # 1.31.0 (2025-12-04)